Some quick changes
Turns out that the national tract data has merged tracts into their
counties. We don't want this, so for now we're back to using state
specific ones.

In order to mesh the coordinate parser more nicely with the tract
parser, we now treat the geoID as a string instead of a long.